backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
scapy==2.6.1
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
pydantic>=2.10.0
websockets==14.1
python-multipart==0.0.19
